+++ src/doc_util/event_builder.h
@@ -77,7 +77,7 @@ static BeatFraction at(int start, int num, int den) {
 
 static Note pitch(int octave, int semitone) {
     return Note(Chromatic(
-        std::clamp(12 * octave + semitone, 0, CHROMATIC_COUNT - 1)
+        std::clamp(12 * octave + semitone, 0, (int) CHROMATIC_COUNT - 1)
     ));
 }
